fix: only autocorrect org names
