+++ cli/src/opts/mod.rs
@@ -2,12 +2,14 @@ pub mod cast;
 pub mod forge;
 
 mod chain;
+mod dependency;
 mod ethereum;
 mod multi_wallet;
 mod transaction;
 mod wallet;
 
 pub use chain::*;
+pub use dependency::*;
 pub use ethereum::*;
 pub use multi_wallet::*;
 pub use transaction::*;
